Split __init__.py into logically separate submodules
Move the implementation out of __init__.py into base (element data,
constants, shared utilities), levelnames (configuration/term parsing),
ionhandlers (ion selection and handler defaults), phixs (cross-section
downsampling and hydrogenic estimates), iondata (per-ion reading),
output (ARTIS file writers), and cli (argument parsing and main loop).
__init__.py becomes a re-export facade, so artisatomic.name access in
the readers, the tests, and the makeartisatomicfiles entry point are
unchanged. Output files are byte-identical for all five test sets.
