+++ artisatomic/__init__.py
@@ -4,32 +4,11 @@
 
 Each data source has its own read*.py module; this package selects a handler per ion, calls it,
 and writes the combined result to adata.txt, transitiondata.txt and phixsdata_v2.txt.
+
+The implementation lives in submodules (base, levelnames, ionhandlers, phixs, iondata, output,
+cli); everything is re-exported here so readers and scripts can keep using the flat
+artisatomic.name interface.
 """
-
-import argparse
-import contextlib
-import glob
-import json
-import math
-import multiprocessing as mp
-import operator
-import os
-import sys
-import typing as t
-from collections.abc import Callable
-from collections.abc import Iterable
-from collections.abc import Sequence
-from functools import lru_cache
-from functools import partial
-from pathlib import Path
-
-import argcomplete
-import numpy as np
-import numpy.typing as npt
-import pandas as pd
-import polars as pl
-from scipy import integrate  # pyright: ignore[reportMissingTypeStubs]
-from scipy import interpolate  # pyright: ignore[reportMissingTypeStubs]
 
 from artisatomic import groundstatesonlynist
 from artisatomic import readboyledata
@@ -41,1358 +20,105 @@ from artisatomic import readkuruczdata
 from artisatomic import readnahardata
 from artisatomic import readqubdata
 from artisatomic import readtanakajpltdata
-
-# import artisatomic.readlisbondata as readlisbondata
-
-PYDIR = Path(__file__).parent.resolve()
-atomicdata = pd.read_csv(PYDIR / "atomic_properties.txt", sep=r"\s+", comment="#")
-atomicdata = atomicdata.apply(lambda x: x.fillna(x.number / 0.45), axis=1)  # estimate unknown atomic mass as Z / 0.45
-elsymbols = ["n", *list(atomicdata["symbol"].values)]
-atomic_weights = ["n", *list(atomicdata["mass"].values)]
-
-roman_numerals = (
-    "",
-    "I",
-    "II",
-    "III",
-    "IV",
-    "V",
-    "VI",
-    "VII",
-    "VIII",
-    "IX",
-    "X",
-    "XI",
-    "XII",
-    "XIII",
-    "XIV",
-    "XV",
-    "XVI",
-    "XVII",
-    "XVIII",
-    "XIX",
-    "XX",
-)
-
-
-def get_ion_handlers() -> list[tuple[int, list[int | tuple[int, str]]]]:
-    """Get the ions to process and the handler to read each one with.
-
-    Read from artisatomicionhandlers.json when that file exists, so a run can be repeated
-    exactly; otherwise built from the hard-coded selection below plus whatever the readers'
-    extend_ion_list() functions find data for.
-    """
-    inputhandlersfile = Path("artisatomicionhandlers.json")
-
-    if inputhandlersfile.exists():
-        print(f"Reading {inputhandlersfile}")
-        with inputhandlersfile.open(encoding="utf-8") as f:
-            return sort_ion_handlers(json.load(f))
-
-    ion_handlers: list[tuple[int, list[int | tuple[int, str]]]] = []
-
-    ion_handlers = [
-        #     (2, [(3, "boyle")]),
-        (38, [(1, "kurucz"), (2, "kurucz"), (3, "kurucz")]),
-        (39, [(1, "kurucz"), (2, "kurucz")]),
-        (40, [(1, "kurucz"), (2, "kurucz"), (3, "kurucz")]),
-    ]
-
-    # include everything we have data for
-    ion_handlers = readqubdata.extend_ion_list(ion_handlers)
-    ion_handlers = readhillierdata.extend_ion_list(ion_handlers, maxionstage=5, include_hydrogen=True)
-    # ion_handlers = readdreamdata.extend_ion_list(ion_handlers)
-    # ion_handlers = readfacdata.extend_ion_list(ion_handlers)
-    ion_handlers = readfloers25data.extend_ion_list(ion_handlers, calibrated=True)
-    ion_handlers = readtanakajpltdata.extend_ion_list(ion_handlers, maxionstage=5)
-    # ion_handlers = groundstatesonlynist.extend_ion_list(ion_handlers)
-
-    return sort_ion_handlers(ion_handlers)
-
-
-USE_QUB_COBALT = False
-
-ryd_to_ev = 13.605693122994232
-hc_in_ev_cm = 0.0001239841984332003
-hc_in_ev_angstrom = 12398.419843320025
-h_in_ev_seconds = 4.135667696923859e-15
-
-
-def split_element_ionstage_str(ionstr: str) -> tuple[int, int]:
-    """Split a string like 'FeII' into (atomic_number, ion_stage).
-
-    Splitting on `ionstr.rstrip("IVX")` destroys the symbols of the elements whose symbols are
-    made only of those letters: V (vanadium) and I (iodine). Instead find the split point where
-    the prefix is an element symbol and the suffix is a Roman numeral. Element symbols have a
-    lowercase second letter and Roman numerals are uppercase, so the match is unambiguous.
-    """
-    for splitpos in range(1, len(ionstr)):
-        elsym, ion_stage_roman = ionstr[:splitpos], ionstr[splitpos:]
-        if elsym in elsymbols and ion_stage_roman in roman_numerals[1:]:
-            return elsymbols.index(elsym), roman_numerals.index(ion_stage_roman)
-
-    msg = f"Could not split '{ionstr}' into an element symbol and a Roman numeral ion stage"
-    raise ValueError(msg)
-
-
-def get_ion_stage(entry: int | tuple[int, str]) -> int:
-    """Ion stage of an ion_handlers entry, which is either a bare ion stage or (ion_stage, handler)."""
-    return entry if isinstance(entry, int) else entry[0]
-
-
-def sort_ion_handlers(
-    ion_handlers: list[tuple[int, list[int | tuple[int, str]]]],
-) -> list[tuple[int, list[int | tuple[int, str]]]]:
-    """Sort by atomic number, and each element's ions by ion stage.
-
-    process_files() relies on ascending ion stages to identify the top ion and to find each ion's
-    photoionisation target, so normalise the order here, before the handler list is written to
-    artisatomicionhandlers.json and passed to write_compositionfile().
-    """
-    return sorted(
-        ((atomic_number, sorted(listions, key=get_ion_stage)) for atomic_number, listions in ion_handlers),
-        key=operator.itemgetter(0),
-    )
-
-
-def drop_handlers(list_ions: list[int | tuple[int, str]]) -> list[int]:
-    """Replace [(ion_stage, 'handler1'), (ion_stage2, 'handler2'), ion_stage3] with [ion_stage1, ion_stage2, ion_stage3]."""
-    return [get_ion_stage(ion_stage) for ion_stage in list_ions]
-
-
-def add_handler_if_not_set(
-    ion_handlers: list[tuple[int, list[int | tuple[int, str]]]],
-    atomic_number: int | str,
-    ion_stage: int | str,
-    handler: str,
-) -> list[tuple[int, list[int | tuple[int, str]]]]:
-    """Return a new ion_handlers list with (ion_stage, handler) added unless the ion is already present.
-
-    The input list is not modified, so the return value must be used.
-    """
-    # readers derive these from pandas/numpy data, and json.dump() in main() cannot serialise
-    # numpy integers, so normalise here rather than in each caller
-    atomic_number = int(atomic_number)
-    ion_stage = int(ion_stage)
-
-    ion_handlers_out: list[tuple[int, list[int | tuple[int, str]]]] = []
-    found_element = False
-    for tmp_atomic_number, list_ions_handlers in ion_handlers:
-        list_ions_handlers_out: list[int | tuple[int, str]] = list(list_ions_handlers)
-        if tmp_atomic_number == atomic_number:
-            found_element = True
-            if ion_stage not in [get_ion_stage(x) for x in list_ions_handlers_out]:
-                # add an ion that is not present in the element's list
-                list_ions_handlers_out.append((ion_stage, handler))
-        ion_handlers_out.append((tmp_atomic_number, list_ions_handlers_out))
-
-    if not found_element:
-        new_element_ions: list[int | tuple[int, str]] = [(ion_stage, handler)]
-        ion_handlers_out.append((atomic_number, new_element_ions))
-
-    return sort_ion_handlers(ion_handlers_out)
-
-
-# The id-keyed transition columns write_transition_data() needs, so an empty frame still carries
-# them. Name-keyed frames get lowerlevel/upperlevel from add_level_ids_forbidden() instead.
-empty_transitions_schema = pl.Schema({"lowerlevel": pl.Int64, "upperlevel": pl.Int64, "A": pl.Float64})
-
-
-def leveltuples_to_pldataframe(energy_levels) -> pl.DataFrame:
-    """Convert a list of level tuples (or a DataFrame) into a DataFrame with a zero-based levelid column.
-
-    Level ids are zero-based everywhere in memory; the 1-based numbering of the output files is
-    applied by the write_*() functions.
-    """
-    dflevels = energy_levels if isinstance(energy_levels, pl.DataFrame) else pl.DataFrame(energy_levels)
-
-    if "levelid" not in dflevels.columns:
-        dflevels = dflevels.with_row_index(name="levelid")
-
-    dflevels = dflevels.with_columns(pl.col("levelid").cast(pl.Int64))
-
-    # the frame is indexed by level id elsewhere, so a reader-supplied levelid must be contiguous
-    # and zero-based. Not an assert: input validation must survive python -O.
-    if not dflevels["levelid"].equals(pl.int_range(dflevels.height, dtype=pl.Int64, eager=True)):
-        msg = "level ids must be contiguous and start at zero"
-        raise ValueError(msg)
-
-    return dflevels
-
-
-def main(args: argparse.Namespace | None = None, argsraw: Sequence[str] | None = None, **kwargs: t.Any) -> None:
-    """Write an ARTIS atomic database from the configured ions and handlers."""
-    if args is None:
-        parser = argparse.ArgumentParser(
-            formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            description="Produce an ARTIS atomic database from published atomic data sets.",
-        )
-        parser.add_argument("-output_folder", action="store", default="artis_files", help="Folder for output files")
-        parser.add_argument(
-            "-output_folder_logs", action="store", default="atomic_data_logs", help="Folder for log files"
-        )
-        parser.add_argument(
-            "-nphixspoints", type=int, default=100, help="Number of cross section points to save in output"
-        )
-        parser.add_argument(
-            "-phixsnuincrement",
-            type=float,
-            default=0.03,
-            help="Fraction of nu_edge incremented for each cross section point",
-        )
-        parser.add_argument(
-            "-optimaltemperature",
-            type=int,
-            default=6000,
-            help=(
-                "(Electron and excitation) temperature at which recombination rate "
-                "should be constant when downsampling cross sections"
-            ),
-        )
-        parser.add_argument(
-            "-electrontemperature",
-            type=int,
-            default=6000,
-            help="Temperature for choosing effective collision strengths",
-        )
-        parser.add_argument(
-            "--nophixs", action="store_true", help="Don't generate cross sections and write to phixsdata_v2.txt file"
-        )
-
-        parser.add_argument(
-            "-nlevels_hydrogenic_for_unknown_phixs",
-            type=int,
-            default=100,
-            help=(
-                "Consider this many of the lowest levels of any ion whose handler supplied no"
-                " cross sections at all, and estimate a hydrogenic one for each, or 0 to disable."
-                " Negative values are rejected. Fewer tables than this can result, because a level"
-                " at or above the ionization energy is skipped but still counts towards the limit."
-                " An ion with even one cross section from its data source is left untouched, so"
-                " this never replaces or extends measured data. Excludes the top ion, which has no"
-                " upper ion to photoionise to."
-            ),
-        )
-
-        parser.set_defaults(**kwargs)
-        argcomplete.autocomplete(parser)
-        args = parser.parse_args(argsraw)
-
-    # 0 is the way to switch the estimate off, so a negative value is a typo rather than a
-    # quieter way of saying the same thing
-    if args.nlevels_hydrogenic_for_unknown_phixs < 0:
-        msg = f"-nlevels_hydrogenic_for_unknown_phixs must not be negative, got {args.nlevels_hydrogenic_for_unknown_phixs}"
-        raise ValueError(msg)
-
-    ion_handlers = get_ion_handlers()
-
-    assert len(ion_handlers) > 0
-    readhillierdata.read_hyd_phixsdata()
-
-    os.makedirs(args.output_folder, exist_ok=True)
-
-    log_folder = Path(args.output_folder) / args.output_folder_logs
-    if log_folder.exists():
-        # delete any existing log files
-        logfiles = glob.glob(os.path.join(log_folder, "*.txt"))
-        for logfile in logfiles:
-            Path(logfile).unlink(missing_ok=True)
-            print("deleting", logfile)
-    else:
-        os.makedirs(log_folder, exist_ok=True)
-
-    with Path(log_folder, "artisatomicionhandlers.json").open("w", encoding="utf-8") as f:
-        json.dump(obj=ion_handlers, fp=f)
-    write_compositionfile(ion_handlers, args)
-    clear_files(args)
-    process_files(ion_handlers, args)
-
-
-def clear_files(args: argparse.Namespace) -> None:
-    """Truncate the output files and write the phixs header, which the ions are appended to."""
-    with (
-        open(os.path.join(args.output_folder, "adata.txt"), "w", encoding="utf-8"),
-        open(os.path.join(args.output_folder, "transitiondata.txt"), "w", encoding="utf-8"),
-        open(os.path.join(args.output_folder, "phixsdata_v2.txt"), "w", encoding="utf-8") as fphixs,
-    ):
-        fphixs.write(f"{args.nphixspoints:d}\n")
-        fphixs.write(f"{args.phixsnuincrement:14.7e}\n")
-
-
-def process_files(ion_handlers: list[tuple[int, list[int | tuple[int, str]]]], args: argparse.Namespace) -> None:
-    """Read every configured ion and append it to the output files, one element at a time.
-
-    Ion stages are processed in ascending order so that each ion's photoionisation targets, which
-    are levels of the next ion up, are already known, and so the top ion can be identified and
-    given no cross sections.
-    """
-    for atomic_number, listions in ion_handlers:
-        if not listions:
-            continue
-
-        iondatalist = [
-            read_ion_data(atomic_number, ion_stage_entry, is_top_ion=(i == len(listions) - 1), args=args)
-            for i, ion_stage_entry in enumerate(listions)
-        ]
-
-        write_output_files(atomic_number, iondatalist, args)
-
-
-class IonData(t.NamedTuple):
-    """Levels, transitions, and photoionisation data for a single ion, read from one of the source datasets."""
-
-    ion_stage: int
-    handler: str
-    ionization_energy_ev: float
-    dfenergylevels: pl.DataFrame
-    dftransitions: pl.DataFrame
-    transition_count_of_level_name: dict[str, int]
-    upsilondict: dict[tuple[int, int], float]
-    # core state id n is at index n - 1 ([] if unavailable)
-    nahar_core_states: list[readnahardata.NaharCoreState]
-    # keys are (2S+1, L, parity, index in symmetry), values are strings of electron configuration
-    nahar_configurations: dict[tuple[int, int, int, int], str]
-    # None where a level has no photoionisation data (and None entirely if none was read)
-    hillier_photoion_targetconfigs: list[list[tuple[str, float]] | None] | None
-    photoionization_crosssections: npt.NDArray[np.float64]  # cross sections in Mb, indexed by level id
-    photoionization_targetfractions: list[list[tuple[int, float]]]  # indexed by level id
-    photoionization_thresholds_ev: npt.NDArray[np.float64]  # indexed by level id
-
-
-def get_default_handler(atomic_number: int, ion_stage: int) -> str:
-    """Get the data source to use for an ion when the handler list does not name one."""
-    if atomic_number == 2 and ion_stage == 3:
-        return "boyle"
-    if USE_QUB_COBALT and atomic_number == 27:
-        return "qub_cobalt"
-    if atomic_number <= 28 or atomic_number == 56:  # Hillier data only
-        return "cmfgen"
-    # a QUB calculation is preferred over the Kurucz line lists, and over DREAM for W, Pt and Au
-    if (atomic_number, ion_stage) in readqubdata.default_handler_ions:
-        return "qub_data"
-    if atomic_number >= 57:  # DREAM database of Z > 57
-        return "dream"
-    return "kurucz"
-
-
-# handlers whose readers share a common call signature and return
-# (ionization_energy_ev, energy_levels, transitions, transition_count_of_level_name)
-# with an additional upsilondict for qub_data
-simple_handler_readers: dict[str, Callable[..., tuple[t.Any, ...]]] = {
-    "boyle": lambda atomic_number, ion_stage, _flog: readboyledata.read_levels_and_transitions(
-        atomic_number, ion_stage
-    ),
-    "kurucz": readkuruczdata.read_levels_and_transitions,
-    "dream": readdreamdata.read_levels_and_transitions,  # DREAM database of Z >= 57
-    # "lisbon": readlisbondata.read_levels_and_transitions,
-    "floers25calib": lambda atomic_number, ion_stage, flog: readfloers25data.read_levels_and_transitions(
-        atomic_number, ion_stage, flog, calibrated=True
-    ),
-    "floers25uncalib": lambda atomic_number, ion_stage, flog: readfloers25data.read_levels_and_transitions(
-        atomic_number, ion_stage, flog, calibrated=False
-    ),
-    "fac": readfacdata.read_levels_and_transitions,  # early version of floers25 calib data
-    "tanakajplt": readtanakajpltdata.read_levels_and_transitions,  # Tanaka Japan-Lithuania database of 26 <= Z <= 88
-    "gsnist": groundstatesonlynist.read_ground_levels,  # ground states taken from NIST
-    "qub_data": readqubdata.read_qub_levels_and_transitions,  # also returns an upsilondict
-}
-
-
-def read_ion_data(
-    atomic_number: int, ion_stage_entry: int | tuple[int, str], is_top_ion: bool, args: argparse.Namespace
-) -> IonData:
-    """Read a single ion's data from its source dataset."""
-    if isinstance(ion_stage_entry, int):
-        ion_stage = ion_stage_entry
-        handler = get_default_handler(atomic_number, ion_stage)
-    else:
-        ion_stage, handler = ion_stage_entry
-
-    ionization_energy_ev = 0.0
-    transition_count_of_level_name: dict[str, int] = {}
-    upsilondict: dict[tuple[int, int], float] = {}
-    nahar_core_states: list[readnahardata.NaharCoreState] = []
-    nahar_configurations: dict[tuple[int, int, int, int], str] = {}
-    hillier_photoion_targetconfigs: list[list[tuple[str, float]] | None] | None = None
-    # empty until a handler below reads photoionisation data (and left empty for the top ion)
-    photoionization_crosssections: npt.NDArray[np.float64] = np.empty((0, args.nphixspoints))  # in Mb
-    photoionization_targetfractions: list[list[tuple[int, float]]] = []
-    photoionization_thresholds_ev: npt.NDArray[np.float64] = np.empty(0)
-
-    logfilepath = Path(
-        args.output_folder, args.output_folder_logs, f"{elsymbols[atomic_number].lower()}{ion_stage:d}.txt"
-    )
-    with logfilepath.open("w", encoding="utf-8") as flog:
-        log_and_print(
-            flog,
-            f"\n===========> Z={atomic_number} {elsymbols[atomic_number]} {roman_numerals[ion_stage]} input:",
-        )
-        log_and_print(flog, f"Source handler: {handler}")
-        if handler == "qub_cobalt":
-            if ion_stage in {3, 4}:  # QUB levels and transitions, or single-level Co IV
-                (
-                    ionization_energy_ev,
-                    energy_levels,
-                    transitions,
-                    transition_count_of_level_name,
-                    upsilondict,
-                ) = readqubdata.read_qub_levels_and_transitions(atomic_number, ion_stage, flog)
-            else:  # hillier levels and transitions
-                # if ion_stage == 2:
-                #     upsilondict = readstoreydata.read_storey_2016_upsilondata(flog)
-                (
-                    ionization_energy_ev,
-                    energy_levels,
-                    transitions,
-                    transition_count_of_level_name,
-                ) = readhillierdata.read_levels_and_transitions(atomic_number, ion_stage, flog)
-
-            if not is_top_ion and not args.nophixs:  # don't get cross sections for top ion
-                (
-                    photoionization_crosssections,
-                    photoionization_targetfractions,
-                    photoionization_thresholds_ev,
-                ) = readqubdata.read_qub_photoionizations(
-                    atomic_number, ion_stage, levelcount=len(energy_levels), args=args, flog=flog
-                )
-
-        elif handler == "nahar":
-            path_nahar_energy_file = f"atomic-data-nahar/{elsymbols[atomic_number].lower()}{ion_stage:d}.en.ls.txt"
-            path_nahar_px_file = f"atomic-data-nahar/{elsymbols[atomic_number].lower()}{ion_stage:d}.ptpx.txt"
-            (
-                nahar_energy_levels,
-                nahar_core_states,
-                nahar_configurations,
-                nahar_ionization_potential_rydberg,
-            ) = readnahardata.read_nahar_energy_level_file(path_nahar_energy_file, atomic_number, ion_stage, flog)
-            ionization_energy_ev = nahar_ionization_potential_rydberg * ryd_to_ev
-
-            # keys are (2S+1, L, parity, indexinsymmetry), values are lists of
-            # (energy in Rydberg, cross section in Mb) tuples
-            nahar_phixs_tables: dict[tuple[int, int, int, int], npt.NDArray[np.float64]] = {}
-            thresholds_ev_dict: dict[tuple[int, int, int, int], float] = {}
-            if not is_top_ion:  # don't get cross sections for top ion
-                log_and_print(flog, f"Reading {path_for_log(path_nahar_px_file)}")
-                nahar_phixs_tables, thresholds_ev_dict = readnahardata.read_nahar_phixs_tables(
-                    path_nahar_px_file, atomic_number, ion_stage, args
-                )
-
-            (energy_levels, photoionization_crosssections, photoionization_thresholds_ev) = (
-                readnahardata.build_nahar_levels_and_phixs(
-                    nahar_energy_levels, nahar_phixs_tables, thresholds_ev_dict, args, flog
-                )
-            )
-            # the Nahar data set gives no bound-bound transitions
-            transitions = pl.DataFrame(schema=empty_transitions_schema)
-
-        elif handler == "cmfgen":  # Hillier CMFGEN data only
-            (
-                ionization_energy_ev,
-                energy_levels,
-                transitions,
-                transition_count_of_level_name,
-            ) = readhillierdata.read_levels_and_transitions(atomic_number, ion_stage, flog)
-
-            if len(upsilondict) == 0:
-                upsilondict = readhillierdata.read_coldata(atomic_number, ion_stage, energy_levels, flog, args)
-
-            if not is_top_ion and not args.nophixs:  # don't get cross sections for top ion
-                (
-                    photoionization_crosssections,
-                    hillier_photoion_targetconfigs,
-                    photoionization_thresholds_ev,
-                ) = readhillierdata.read_phixs_tables(atomic_number, ion_stage, energy_levels, args, flog)
-            else:
-                hillier_photoion_targetconfigs = None
-
-        elif handler in simple_handler_readers:
-            result = simple_handler_readers[handler](atomic_number, ion_stage, flog)
-            if len(result) == 5:
-                (ionization_energy_ev, energy_levels, transitions, transition_count_of_level_name, upsilondict) = result
-            else:
-                (ionization_energy_ev, energy_levels, transitions, transition_count_of_level_name) = result
-
-        else:
-            raise ValueError(f"Unknown handler: {handler}")
-
-    dfenergylevels = leveltuples_to_pldataframe(energy_levels)
-
-    # the len() == 0 test is what limits the estimate to ions the handler gave nothing for: an ion
-    # with even one cross-section table is left alone, so measured data is never replaced. The top
-    # ion is excluded because there is no upper ion for it to photoionise to.
-    if (
-        not is_top_ion
-        and not args.nophixs
-        and len(photoionization_crosssections) == 0
-        and args.nlevels_hydrogenic_for_unknown_phixs > 0
-    ):
-        (
-            photoionization_crosssections,
-            photoionization_targetfractions,
-            photoionization_thresholds_ev,
-        ) = match_hydrogenic_phixs(atomic_number, dfenergylevels, ionization_energy_ev, handler, args)
-
-    dftransitions = transitions if isinstance(transitions, pl.DataFrame) else pl.DataFrame(transitions)
-
-    return IonData(
-        ion_stage=ion_stage,
-        handler=handler,
-        ionization_energy_ev=ionization_energy_ev,
-        dfenergylevels=dfenergylevels,
-        dftransitions=dftransitions,
-        transition_count_of_level_name=transition_count_of_level_name,
-        upsilondict=upsilondict,
-        nahar_core_states=nahar_core_states,
-        nahar_configurations=nahar_configurations,
-        hillier_photoion_targetconfigs=hillier_photoion_targetconfigs,
-        photoionization_crosssections=photoionization_crosssections,
-        photoionization_targetfractions=photoionization_targetfractions,
-        photoionization_thresholds_ev=photoionization_thresholds_ev,
-    )
-
-
-def log_and_print(flog, strout):
-    """Write a line to both stdout and this ion's log file."""
-    print(strout)
-    flog.write(strout + "\n")
-
-
-def path_for_log(filepath: str | Path) -> str:
-    """Render an input data path relative to the repository root where possible.
-
-    The log files are compared by checksum in CI, so an absolute path would make them depend on
-    where the repository happens to be checked out. Paths outside the repository (some readers
-    load data from elsewhere) are returned unchanged.
-    """
-    try:
-        return str(Path(filepath).resolve().relative_to(PYDIR.parent))
-    except ValueError:
-        return str(filepath)
-
-
-def isfloat(value: t.Any) -> bool:
-    """Whether a string parses as a float, accepting Fortran's D exponent (1.5D-3)."""
-    try:
-        float(value.replace("D", "E"))
-    except ValueError:
-        return False
-
-    return True
-
-
-compression_extensions = ("", ".zst", ".gz", ".xz")
-
-
-def find_file_check_extension(filename: str | Path) -> Path | None:
-    """Find a data file by its plain name, accepting any of the compressed variants of that name.
-
-    Returns None if neither the plain name nor any compressed form exists, so that callers which
-    treat a missing file as "no data for this ion" can say so without opening it.
-    """
-    return next((path for ext in compression_extensions if (path := Path(f"{filename}{ext}")).is_file()), None)
-
-
-def xopen_check_extension(filename: str | Path, **kwargs: t.Any) -> t.IO[t.Any]:
-    """Open a data file, trying the compressed variants of the name if it does not exist.
-
-    The data sets ship some files compressed and some not, and which ones varies between
-    downloads, so callers name the plain file and this finds whichever form is present.
-    """
-    from xopen import xopen
-
-    filepath = find_file_check_extension(filename)
-    if filepath is None:
-        filepaths = [f"{filename}{ext}" for ext in compression_extensions]
-        msg = f"Could not find any of the following files:\n  {'\n  '.join(filepaths)}."
-        raise FileNotFoundError(msg)
-
-    return xopen(filepath, **kwargs)
-
-
-# split a list into evenly sized chunks
-def chunks[T](listin: list[T], chunk_size: int) -> list[list[T]]:
-    """Split a list into consecutive chunks of at most chunk_size items."""
-    return [listin[i : i + chunk_size] for i in range(0, len(listin), chunk_size)]
-
-
-@lru_cache(maxsize=1)
-def get_nist_ionization_energies_ev() -> dict[tuple[int, int], float]:
-    """Get a dictionary where dictioniz[(atomic_number, ion_sage)] = ionization_energy_ev."""
-    dfnist = pd.read_csv(
-        PYDIR / "nist_ionization.txt",
-        sep="\t",
-        usecols=["At. num", "Ion Charge", "Ionization Energy (a) (eV)"],
-    )
-
-    dictioniz = {}
-    for atomic_number, ion_charge, ioniz_ev in dfnist[
-        ["At. num", "Ion Charge", "Ionization Energy (a) (eV)"]
-    ].itertuples(index=False):
-        with contextlib.suppress(ValueError):
-            ion_stage = int(ion_charge) + 1
-            dictioniz[int(atomic_number), ion_stage] = ioniz_ev
-    return dictioniz
-
-
-def match_hydrogenic_phixs(
-    atomic_number: int, energy_levels: pl.DataFrame, ionization_energy_ev: float, ion_handler: str, args
-) -> tuple[npt.NDArray[np.float64], list[list[tuple[int, float]]], npt.NDArray[np.float64]]:
-    """Estimate photoionization cross sections for a data set that supplies none.
-
-    Applies to any handler, not just one source: a hydrogenic cross section is assigned to each of
-    the lowest -nlevels_hydrogenic_for_unknown_phixs levels, scaled to that level's own ionisation
-    threshold, with the upper ion's ground state as the only target. That option defaults to 100,
-    so this is on unless it is set to 0. It bounds the levels considered rather than the tables
-    produced: a level at or above the ionization energy is skipped but still counts towards it.
-
-    The caller only reaches this for an ion whose handler returned no cross sections at all, so
-    real data is never replaced or extended by an estimate. The granularity is the whole ion: an
-    ion whose handler covered even one level keeps exactly the levels that handler covered, and
-    the rest are left without photoionization rather than filled in hydrogenically.
-    """
-    dict_get_n_func = {
-        "tanakajplt": readtanakajpltdata.get_level_valence_n,
-        "kurucz": readkuruczdata.get_level_valence_n,
-        "fac": readfacdata.get_level_valence_n,
-        "floers25calib": readfloers25data.get_level_valence_n,
-        "floers25uncalib": readfloers25data.get_level_valence_n,
-        "qub_data": readqubdata.get_level_valence_n,
-    }
-    if ion_handler not in dict_get_n_func:
-        print(
-            f"WARNING: Can't assign hydrogenic photoionization cross sections because I don't know how to find principle quantum numbers for {ion_handler} levels"
-        )
-        return np.empty((0, args.nphixspoints)), [], np.empty(0)
-
-    get_n = dict_get_n_func[ion_handler]
-    print(f"using hydrogenic photoionization cross sections for Z={atomic_number} {elsymbols[atomic_number]}")
-
-    photoionization_crosssections = np.zeros((energy_levels.height, args.nphixspoints))
-    photoionization_targetfractions: list[list[tuple[int, float]]] = [[] for _ in range(energy_levels.height)]
-    photoionization_thresholds_ev = np.full(energy_levels.height, np.nan)
-    phixstables = {}
-    for levelindex, level in enumerate(energy_levels.iter_rows(named=True)):
-        if levelindex >= args.nlevels_hydrogenic_for_unknown_phixs:
-            break
-        en_ev = hc_in_ev_cm * level["energyabovegsinpercm"]
-        threshold_ev = ionization_energy_ev - en_ev
-        if threshold_ev <= 0.0:
-            # level lies above the ionization energy, so there is nothing to ionize from
-            continue
-        photoionization_thresholds_ev[levelindex] = threshold_ev
-        lambda_angstrom = hc_in_ev_angstrom / threshold_ev
-
-        n = get_n(level["levelname"])
-        # get_hydrogenic_n_phixstable() already scales by the effective charge, since its
-        # scale factor 7.91 / (E_threshold / Ryd) / n is the Kramers result 7.91 * n / Z_eff^2
-        phixstables[levelindex] = readhillierdata.get_hydrogenic_n_phixstable(lambda_angstrom=lambda_angstrom, n=n)
-        photoionization_targetfractions[levelindex] = [(0, 1.0)]  # the upper ion's ground state
-
-    reduced_phixs_dict = reduce_phixs_tables(
-        phixstables, args.optimaltemperature, args.nphixspoints, args.phixsnuincrement
-    )
-    for levelindex, reduced_phixs_table in reduced_phixs_dict.items():
-        photoionization_crosssections[levelindex] = reduced_phixs_table
-
-    return photoionization_crosssections, photoionization_targetfractions, photoionization_thresholds_ev
-
-
-def parallel_map[ResultType](
-    fn: Callable[..., ResultType],
-    *iterables: Iterable[t.Any],
-    **kwargs: t.Any,
-) -> list[ResultType]:
-    """Execute a parallel map with a progress bar using either multithreading (for free-threading python) or multiprocessing."""
-    # use a thread pool if we have no GIL (free threading)
-    use_multiprocessing = sys._is_gil_enabled()  # ruff: ignore[private-member-access]
-
-    if use_multiprocessing:
-        mp.set_start_method("spawn", force=True)
-        from tqdm.contrib.concurrent import process_map
-
-        results = process_map(fn, *iterables, **kwargs)  # type: ignore[arg-type] # zuban: ignore[no-untyped-call]
-    else:
-        from tqdm.contrib.concurrent import thread_map
-
-        results = thread_map(fn, *iterables, **kwargs)  # type: ignore[arg-type] # zuban: ignore[no-untyped-call]
-
-    assert isinstance(results, list)
-    return results
-
-
-def reduce_phixs_tables[KeyType](
-    dicttables: dict[KeyType, npt.NDArray[np.float64]],
-    optimaltemperature: float,
-    nphixspoints: int,
-    phixsnuincrement: float,
-) -> dict[KeyType, npt.NDArray[np.float64]]:
-    """Downsample each 2D table of (energy, cross section) points into a 1D array.
-
-    Units don't matter, but the first (lowest) energy point is assumed to be the threshold energy
-
-    The key type is preserved: callers index the tables by level name, by Nahar state tuple, or
-    by level id.
-    """
-    print(f"Processing {len(dicttables.keys()):d} phixs tables")
-
-    dictout = dict(
-        zip(
-            dicttables.keys(),
-            parallel_map(
-                partial(
-                    reduce_phixs_tables_worker,
-                    optimaltemperature,
-                    nphixspoints,
-                    phixsnuincrement,
-                ),
-                dicttables.values(),
-            ),
-            strict=True,
-        )
-    )
-
-    return dictout
-
-
-# this method downsamples the photoionization cross section table to a
-# regular grid while keeping the recombination rate integral constant
-# (assuming that the temperature matches)
-def reduce_phixs_tables_worker(
-    optimaltemperature: float,
-    nphixspoints: int,
-    phixsnuincrement: float,
-    tablein: np.ndarray,
-) -> np.ndarray:
-    """Downsample one cross-section table onto the output's nu/nu_edge grid.
-
-    Each output point is the average of the input over that point's frequency bin, weighted by
-    nu^2 exp(-h nu / k T) so that the recombination rate at optimaltemperature is preserved
-    rather than the cross section itself.
-    """
-    ryd_to_hz = 3289841960250880.5
-    h_over_kb_in_K_sec = 4.799243073366221e-11
-
-    # proportional to recombination rate
-    # nu0 = 1e16
-    # fac = math.exp(h_over_kb_in_K_sec * nu0 / optimaltemperature)
-
-    def integrand(nu):
-        """Weight for averaging the cross section: proportional to the recombination rate."""
-        return (nu**2) * math.exp(-h_over_kb_in_K_sec * nu / optimaltemperature)
-
-    # def integrand_vec(nu_list):
-    #    return [(nu ** 2) * math.exp(- h_over_kb_in_K_sec * (nu - nu0) / optimaltemperature)
-    #            for nu in nu_list]
-
-    integrand_vec = np.vectorize(integrand)
-
-    xgrid = np.linspace(1.0, 1.0 + phixsnuincrement * (nphixspoints + 1), num=nphixspoints + 1, endpoint=False)
-
-    # for key in keylist:
-    #   tablein = dicttables[key]
-    # # filter zero points out of the table
-    # firstnonzeroindex = 0
-    # for i, point in enumerate(tablein):
-    #     if point[1] != 0.:
-    #         firstnonzeroindex = i
-    #         break
-    # if firstnonzeroindex != 0:
-    #     tablein = tablein[firstnonzeroindex:]
-
-    # table says zero threshold, so avoid divide by zero
-    if tablein[0][0] == 0.0:
-        return np.zeros(nphixspoints)
-
-    threshold_old_ryd = tablein[0][0]
-    # tablein is an array of pairs (energy, phixs cross section)
-
-    # nu0 = tablein[0][0] * ryd_to_hz
-
-    arr_sigma_out = np.empty(nphixspoints)
-    # x is nu/nu_edge
-
-    sigma_interp = interpolate.interp1d(tablein[:, 0], tablein[:, 1], kind="linear", assume_sorted=True)
-
-    for i, _ in enumerate(xgrid[:-1]):
-        iprevious = max(i - 1, 0)
-        enlow = 0.5 * (xgrid[iprevious] + xgrid[i]) * threshold_old_ryd
-        enhigh = 0.5 * (xgrid[i] + xgrid[i + 1]) * threshold_old_ryd
-
-        # start of interval interpolated point, Nahar points, and end of interval interpolated point
-        samples_in_interval = tablein[(enlow <= tablein[:, 0]) & (tablein[:, 0] <= enhigh)]
-
-        if len(samples_in_interval) == 0 or ((samples_in_interval[0, 0] - enlow) / enlow) > 1e-20:
-            if i == 0 and len(samples_in_interval) != 0:
-                print(
-                    f"adding first point {enlow:.4e} {samples_in_interval[0, 0]:.4e} {(samples_in_interval[0, 0] - enlow) / enlow:.4e}"
-                )
-            if enlow <= tablein[-1][0]:
-                new_crosssection = sigma_interp(enlow)
-                if new_crosssection < 0:
-                    print("negative extrap")
-            else:
-                # assume power law decay after last point
-                new_crosssection = tablein[-1][1] * (tablein[-1][0] / enlow) ** 3
-            samples_in_interval = np.vstack([[enlow, new_crosssection], samples_in_interval])
-
-        if (
-            len(samples_in_interval) == 0
-            or ((enhigh - samples_in_interval[-1, 0]) / samples_in_interval[-1, 0]) > 1e-20
-        ):
-            if enhigh <= tablein[-1][0]:
-                new_crosssection = sigma_interp(enhigh)
-                if new_crosssection < 0:
-                    print("negative extrap")
-            else:
-                new_crosssection = (
-                    tablein[-1][1] * (tablein[-1][0] / enhigh) ** 3
-                )  # assume power law decay after last point
-
-            samples_in_interval = np.vstack([samples_in_interval, [enhigh, new_crosssection]])
-
-        nsamples = len(samples_in_interval)
-
-        # integralnosigma, err = integrate.fixed_quad(integrand_vec, enlow, enhigh, n=250)
-        # integralwithsigma, err = integrate.fixed_quad(
-        #    lambda x: sigma_interp(x) * integrand_vec(x), enlow, enhigh, n=250)
-
-        # this is incredibly fast, but maybe not accurate
-        # integralnosigma, err = integrate.quad(integrand, enlow, enhigh, epsrel=1e-2)
-        # integralwithsigma, err = integrate.quad(
-        #    lambda x: sigma_interp(x) * integrand(x), enlow, enhigh, epsrel=1e-2)
-
-        if nsamples >= 50 or enlow > tablein[-1][0]:
-            arr_energyryd = samples_in_interval[:, 0]
-            arr_sigma_megabarns = samples_in_interval[:, 1]
-        else:
-            nsteps = 50  # was 500
-            arr_energyryd = np.linspace(enlow, enhigh, num=nsteps, endpoint=False)
-            arr_sigma_megabarns = np.interp(arr_energyryd, tablein[:, 0], tablein[:, 1])
-        assert isinstance(arr_sigma_megabarns, np.ndarray)
-
-        integrand_vals = integrand_vec(arr_energyryd * ryd_to_hz)
-        if np.any(integrand_vals):
-            sigma_integrand_vals = [
-                sigma * integrand_val for sigma, integrand_val in zip(arr_sigma_megabarns, integrand_vals, strict=True)
-            ]
-
-            integralnosigma = integrate.trapezoid(integrand_vals, arr_energyryd)
-            integralwithsigma = integrate.trapezoid(sigma_integrand_vals, arr_energyryd)
-
-        else:
-            integralnosigma = 1.0
-            integralwithsigma = np.average(arr_sigma_megabarns)
-
-        if integralwithsigma > 0 and integralnosigma > 0:
-            arr_sigma_out[i] = integralwithsigma / integralnosigma
-        elif integralwithsigma == 0:
-            arr_sigma_out[i] = 0.0
-        else:
-            print("Math error: ", i, nsamples, integralwithsigma, integralnosigma)
-            print(samples_in_interval)
-            arr_sigma_out[i] = 0.0
-            # sys.exit()
-
-    return arr_sigma_out
-
-
-alphabets = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ "
-reversedalphabets = "zyxwvutsrqponmlkjihgfedcbaZYXWVUTSRQPONMLKJIHGFEDCBA "
-lchars = "SPDFGHIKLMNOPQRSTUVWXYZ"
-
-
-def interpret_configuration(instr_orig: str) -> tuple[list[str], int, int, int, int]:
-    """Split a level name into its orbitals and term.
-
-    Returns (orbitals, 2S+1, L, parity, index in symmetry), where orbitals is the configuration
-    split into occupied orbitals and parent terms (kept in parentheses), and the index in
-    symmetry comes from the seniority letter if the name has one. Term components that cannot be
-    read come back as -1.
-    """
-    max_n = 20  # maximum possible principle quantum number n
-    instr = instr_orig
-    instr = instr.split("[")[0]  # remove trailing bracketed J value
-
-    if instr[-1] in lchars:
-        term_parity = 0  # even
-    else:
-        term_parity = [0, 1][(instr[-1] == "o")]
-        if all(char not in lchars for char in instr):
-            # This will be an incorrectly formatted QUB file with no term
-            print("Warning: Check QUB file formatting")
-        else:
-            # Preserve previous behaviour
-            instr = instr[:-1]
-
-    term_twosplusone = -1
-    term_l = -1
-    indexinsymmetry = -1
-
-    while instr:
-        if instr[-1] in lchars:
-            term_l = lchars.index(instr[-1])
-            instr = instr[:-1]
-            break
-        if not str.isdigit(instr[-1]):
-            term_parity += (
-                2  # this accounts for things like '3d7(4F)6d_5Pbe' in the Hillier levels. Shouldn't match these
-            )
-        instr = instr[:-1]
-        if all(char not in lchars for char in instr):
-            print("Warning: Check QUB file formatting")
-            break
-
-    if instr and str.isdigit(instr[-1]):
-        term_twosplusone = int(instr[-1])
-        instr = instr[:-1]
-
-    if not instr:
-        pass
-    elif instr[-1] == "_":
-        instr = instr[:-1]
-    elif instr[-1] in alphabets and (
-        (len(instr) < 2 or not str.isdigit(instr[-2])) or (len(instr) < 3 or instr[-3] in lchars.lower())
-    ):
-        # to catch e.g., '3d6(5D)6d4Ge[9/2]' occupation piece 6d, not index d
-        # and 3d7b2Fe is at index b, (keep it from conflicting into the orbital occupation)
-        indexinsymmetry = reversedalphabets.index(instr[-1]) + 1 if term_parity == 1 else alphabets.index(instr[-1]) + 1
-        instr = instr[:-1]
-
-    def is_two_digit_n(strn: str) -> bool:
-        """Test whether two digits are a principal quantum number, not one digit of something else.
-
-        n is written 10 to 19 when it takes two digits, so a leading zero rules it out: the '0' of
-        '3d104s' belongs to the occupation number of the 3d shell, giving 4s and not 04s.
-        """
-        return strn.isdigit() and 10 <= int(strn) < max_n
-
-    electron_config: list[str] = []
-    if not instr.startswith("Eqv st"):
-        while instr:
-            if instr[-1].upper() in lchars:
-                # Orbital with no occupation number, e.g. the '10d' of '3d6(5D)10d_5Pe'. A
-                # digit-letter-letter run keeps its digit, so '4sp(3P)_7Po[2]' gives '4sp'.
-                startpos = (
-                    -3
-                    if len(instr) >= 3
-                    and (is_two_digit_n(instr[-3:-1]) or (instr[-3].isdigit() and not instr[-2].isdigit()))
-                    else -2
-                )
-
-                electron_config.insert(0, instr[startpos:])
-                instr = instr[:startpos]
-            elif instr[-1] == ")":
-                left_bracket_pos = instr.rfind("(")
-                str_parent_term = instr[left_bracket_pos:].replace(" ", "")
-                electron_config.insert(0, str_parent_term)
-                instr = instr[:left_bracket_pos]
-            elif str.isdigit(instr[-1]):  # the number of electrons in an orbital
-                if len(instr) >= 2 and instr[-2].upper() in lchars:
-                    # Single-digit occupation. A two-digit n survives ('10d1') only where the
-                    # digits cannot belong to a preceding orbital: '3d14s2' is ambiguous
-                    # (3d1 4s2 or 3d 14s2), and there the occupation-1 reading wins.
-                    two_digit_n = (
-                        len(instr) >= 4
-                        and is_two_digit_n(instr[-4:-2])
-                        and (len(instr) == 4 or not (instr[-5].isdigit() or instr[-5].upper() in lchars))
-                    )
-                    startpos = -4 if two_digit_n else -3
-                    electron_config.insert(0, instr[startpos:])
-                    instr = instr[:startpos]
-                elif len(instr) >= 3 and str.isdigit(instr[-2]) and instr[-3].upper() in lchars:
-                    # Two-digit occupation, e.g. the closed shells '3d10' and '4f14'. This is
-                    # unambiguous: trailing digits after the orbital letter are the occupation.
-                    startpos = -4 if len(instr) >= 4 and str.isdigit(instr[-4]) else -3
-                    electron_config.insert(0, instr[startpos:])
-                    instr = instr[:startpos]
-                else:
-                    # print('Unknown character ' + instr[-1])
-                    instr = instr[:-1]
-            elif instr[-1] in {"_", " "}:
-                instr = instr[:-1]
-            else:
-                # print('Unknown character ' + instr[-1])
-                instr = instr[:-1]
-
-    # return '{0} {1}{2}{3} index {4}'.format(electron_config, term_twosplusone, lchars[term_l], ['e', 'o'][term_parity], indexinsymmetry)
-    return electron_config, term_twosplusone, term_l, term_parity, indexinsymmetry
-
-
-def get_parity_from_config(instr) -> int:
-    """Parity of a level from its configuration: the sum of l over the occupied orbitals, mod 2.
-
-    Returns 0 for even and 1 for odd. Parent terms in parentheses are not occupied orbitals and
-    are skipped, as are merge markers (see the l >= n check below).
-    """
-    configsplit = interpret_configuration(instr)[0]
-    lchars_lower = lchars.lower()
-    lsum = 0
-    for orbitalstr in configsplit:
-        if orbitalstr.startswith("("):
-            continue  # a parent term such as '(5D)', not an occupied orbital
-        # the orbital letter is the first non-digit, allowing a two-digit principal quantum number
-        lpos = next((pos for pos, char in enumerate(orbitalstr) if char in lchars_lower), None)
-        if lpos is None:
-            # don't fail silently: a skipped orbital means the parity (and hence the forbidden
-            # flags of every transition involving this level) could come out wrong
-            print(f"WARNING: could not read an orbital from '{orbitalstr}' in '{instr}', skipping it for the parity")
-            continue
-        l = lchars_lower.index(orbitalstr[lpos])
-        nelec = int(orbitalstr[lpos + 1 :]) if len(orbitalstr[lpos + 1 :]) > 0 else 1
-
-        # An orbital must satisfy l <= n - 1. CMFGEN's merged high-l levels ('2s2_13w_2W',
-        # '2s2_2p3(4So)5z_5Z') fail this: the letter is a merge marker spanning several l of both
-        # parities, so it has no definite parity and only the real orbitals count.
-        principalquantumnumber = int(orbitalstr[:lpos]) if orbitalstr[:lpos].isdigit() else 0
-        if l >= principalquantumnumber:
-            continue
-
-        lsum += l * nelec
-
-    return lsum % 2
-
-
-def add_level_ids_forbidden(dfenergylevels_ion: pl.DataFrame, dftransitions_ion: pl.DataFrame) -> pl.DataFrame:
-    """Fill in whichever of lowerlevel, upperlevel and forbidden a reader did not supply.
-
-    Readers that key their transitions by level name (namefrom/nameto) get the level ids joined
-    on here; readers that already supply ids keep them. A transition is forbidden when its two
-    levels have the same parity.
-    """
-    if dftransitions_ion.is_empty():
-        return dftransitions_ion
-
-    if "upperlevel" not in dftransitions_ion.columns:
-        dftransitions_ion = dftransitions_ion.join(
-            dfenergylevels_ion.select(pl.col("levelid").alias("upperlevel"), pl.col("levelname").alias("nameto")),
-            on="nameto",
-        )
-
-    if "lowerlevel" not in dftransitions_ion.columns:
-        dftransitions_ion = dftransitions_ion.join(
-            dfenergylevels_ion.select(pl.col("levelid").alias("lowerlevel"), pl.col("levelname").alias("namefrom")),
-            on="namefrom",
-        )
-
-    if "forbidden" not in dftransitions_ion.columns:
-        dftransitions_ion = (
-            dftransitions_ion.join(
-                dfenergylevels_ion.select(
-                    pl.col("levelid").alias("lowerlevel"), pl.col("parity").alias("lower_parity")
-                ),
-                on="lowerlevel",
-            )
-            .join(
-                dfenergylevels_ion.select(
-                    pl.col("levelid").alias("upperlevel"), pl.col("parity").alias("upper_parity")
-                ),
-                on="upperlevel",
-            )
-            .with_columns(forbidden=pl.col("lower_parity") == pl.col("upper_parity"))
-        )
-    return dftransitions_ion
-
-
-def write_output_files(atomic_number: int, iondatalist: list[IonData], args: argparse.Namespace) -> None:
-    """Append one element's ions to adata.txt, transitiondata.txt and phixsdata_v2.txt.
-
-    Takes the element's ions together because an ion's photoionisation targets are levels of the
-    next ion up, so the target fractions can only be resolved once both are in hand.
-    """
-    for i, iondata in enumerate(iondatalist):
-        ion_stage = iondata.ion_stage
-        upsilondict = iondata.upsilondict
-        transition_count_of_level_name = iondata.transition_count_of_level_name
-        ionstr = f"{elsymbols[atomic_number]} {roman_numerals[ion_stage]}"
-
-        flog = open(
-            os.path.join(
-                args.output_folder, args.output_folder_logs, f"{elsymbols[atomic_number].lower()}{ion_stage:d}.txt"
-            ),
-            "a",
-            encoding="utf-8",
-        )
-
-        log_and_print(flog, f"\n===========> Z={atomic_number} {ionstr} output:")
-
-        dfenergylevels_ion = iondata.dfenergylevels
-        dftransitions_ion = iondata.dftransitions
-
-        if dftransitions_ion.is_empty():
-            unused_upsilon_transitions = set()
-        else:
-            dftransitions_ion = add_level_ids_forbidden(dfenergylevels_ion, dftransitions_ion)
-            unused_upsilon_transitions = set(upsilondict.keys()).difference(
-                dftransitions_ion[["lowerlevel", "upperlevel"]].iter_rows(named=False)
-            )
-
-        log_and_print(flog, f"Adding in {len(unused_upsilon_transitions):d} extra transitions with only upsilon values")
-
-        if unused_upsilon_transitions:
-            dfupsilon_only_transitions = pl.DataFrame(
-                list(unused_upsilon_transitions),
-                schema=(("lowerlevel", pl.Int64), ("upperlevel", pl.Int64)),
-                orient="row",
-            ).with_columns(A=0.0)
-            for id_lower, id_upper in dfupsilon_only_transitions[["lowerlevel", "upperlevel"]].iter_rows(named=False):
-                namefrom = dfenergylevels_ion["levelname"][id_upper]
-                nameto = dfenergylevels_ion["levelname"][id_lower]
-
-                transition_count_of_level_name[namefrom] += 1
-                transition_count_of_level_name[nameto] += 1
-
-            dfupsilon_only_transitions = add_level_ids_forbidden(dfenergylevels_ion, dfupsilon_only_transitions)
-            dftransitions_ion = pl.concat([dftransitions_ion, dfupsilon_only_transitions], how="diagonal_relaxed")
-
-        if not dftransitions_ion.is_empty():
-            dftransitions_ion = dftransitions_ion.with_columns(
-                pl.struct(["lowerlevel", "upperlevel", "forbidden"])
-                .map_elements(
-                    lambda row, upsilondict=upsilondict: upsilondict.get(  # type: ignore[misc]
-                        (row["lowerlevel"], row["upperlevel"]),
-                        -2.0 if row["forbidden"] else -1.0,
-                    ),
-                    return_dtype=pl.Float64,
-                )
-                .alias("coll_str")
-            )
-
-        with open(os.path.join(args.output_folder, "adata.txt"), "a", encoding="utf-8") as fatommodels:
-            write_adata(
-                fatommodels,
-                atomic_number,
-                ion_stage,
-                dfenergylevels_ion,
-                iondata.ionization_energy_ev,
-                transition_count_of_level_name,
-                flog,
-            )
-
-        dftransitions_ion = (
-            dftransitions_ion
-            if dftransitions_ion.is_empty()
-            else dftransitions_ion.sort(by=("lowerlevel", "upperlevel"))
-        )
-        with open(os.path.join(args.output_folder, "transitiondata.txt"), "a", encoding="utf-8") as ftransitiondata:
-            write_transition_data(
-                ftransitiondata,
-                atomic_number,
-                ion_stage,
-                dftransitions_ion,
-                flog,
-            )
-
-        if i < len(iondatalist) - 1 and not args.nophixs:  # ignore the top ion
-            photoionization_targetfractions = iondata.photoionization_targetfractions
-            if len(photoionization_targetfractions) < 1:
-                if len(iondata.nahar_core_states) > 0:
-                    photoionization_targetfractions = readnahardata.get_photoiontargetfractions(
-                        dfenergylevels_ion,
-                        iondatalist[i + 1].dfenergylevels,
-                        iondata.nahar_core_states,
-                        iondatalist[i + 1].nahar_configurations,
-                        flog,
-                    )
-                else:
-                    photoionization_targetfractions = readhillierdata.get_photoiontargetfractions(
-                        dfenergylevels_ion, iondatalist[i + 1].dfenergylevels, iondata.hillier_photoion_targetconfigs
-                    )
-
-            with open(os.path.join(args.output_folder, "phixsdata_v2.txt"), "a", encoding="utf-8") as fphixs:
-                write_phixs_data(
-                    fphixs,
-                    atomic_number,
-                    ion_stage,
-                    iondata.photoionization_crosssections,
-                    photoionization_targetfractions,
-                    iondata.photoionization_thresholds_ev,
-                    args,
-                    flog,
-                )
-
-        flog.close()
-
-
-def write_adata(
-    fatommodels,
-    atomic_number: int,
-    ion_stage: int,
-    dfenergylevels: pl.DataFrame,
-    ionization_energy: float,
-    transition_count_of_level_name,
-    flog,
-) -> None:
-    """Append one ion's level list to adata.txt.
-
-    Level ids are zero-based in memory but numbered from one in the output. Each level line ends
-    with the level's name as a free-text comment; artistools reads it back as everything after the
-    fourth field, so it must not be padded.
-    """
-    log_and_print(flog, f"Writing {dfenergylevels.height} levels to 'adata.txt'")
-    fatommodels.write(f"{atomic_number:12d}{ion_stage:12d}{dfenergylevels.height:12d}{ionization_energy:15.7f}\n")
-
-    # every reader names its own levels, and that name is the whole level comment
-    for energylevel in dfenergylevels.iter_rows(named=True):
-        levelname = energylevel.get("levelname", "")
-        transitioncount = transition_count_of_level_name.get(levelname, 0)
-
-        # level ids are zero-based in memory, but the output format numbers them from one
-        fatommodels.write(
-            f"{energylevel['levelid'] + 1:5d} {hc_in_ev_cm * float(energylevel['energyabovegsinpercm']):19.16f} {float(energylevel['g']):8.3f} {transitioncount:4d} {levelname:}\n"
-        )
-
-    fatommodels.write("\n")
-
-
-def write_transition_data(
-    ftransitiondata,
-    atomic_number: int,
-    ion_stage: int,
-    dftransitions_ion: pl.DataFrame,
-    flog,
-) -> None:
-    """Append one ion's transitions to transitiondata.txt.
-
-    Level ids are zero-based in memory but numbered from one in the output, and every transition
-    is written with the lower id first.
-    """
-    log_and_print(flog, f"Writing {dftransitions_ion.height} transitions to 'transitiondata.txt'")
-
-    ftransitiondata.write(f"{atomic_number:7d}{ion_stage:7d}{dftransitions_ion.height:12d}\n")
-
-    if not dftransitions_ion.is_empty():
-        for levelid_lower, levelid_upper, A, coll_str, forbidden in dftransitions_ion[
-            ["lowerlevel", "upperlevel", "A", "coll_str", "forbidden"]
-        ].iter_rows():
-            assert levelid_lower < levelid_upper
-
-            # level ids are zero-based in memory, but the output format numbers them from one
-            ftransitiondata.write(
-                f"{levelid_lower + 1:4d} {levelid_upper + 1:4d} {float(A):11.5e} {coll_str:9.2e} {forbidden:d}\n"
-            )
-
-    ftransitiondata.write("\n")
-
-    num_forbidden_transitions = (
-        0 if dftransitions_ion.is_empty() else dftransitions_ion.filter(pl.col("forbidden")).height
-    )
-
-    num_collision_strengths_applied = (
-        0 if dftransitions_ion.is_empty() else dftransitions_ion.filter(pl.col("coll_str") > 0).height
-    )
-
-    log_and_print(
-        flog,
-        f"  output {dftransitions_ion.height:d} transitions of which {num_forbidden_transitions:d} are forbidden and"
-        f" {num_collision_strengths_applied:d} have collision strengths",
-    )
-
-
-def write_phixs_data(
-    fphixs,
-    atomic_number: int,
-    ion_stage: int,
-    photoionization_crosssections: npt.NDArray[np.float64],
-    photoionization_targetfractions: list[list[tuple[int, float]]],
-    photoionization_thresholds_ev: npt.NDArray[np.float64],
-    args,
-    flog,
-) -> None:
-    """Append one ion's photoionization cross sections to phixsdata_v2.txt.
-
-    Only levels with both targets and a threshold energy are written; the rest are counted and
-    reported. Level ids, of this ion and of the upper ion's targets, are zero-based in memory but
-    numbered from one in the output.
-    """
-    # a level gets a table only if it has targets and a threshold energy. The threshold arrays
-    # start as NaN and are filled in only for levels that got a table, so NaN means "no data".
-    # A negative value is written deliberately (readqubdata): ARTIS reads it as "no value given"
-    # and takes the threshold from the difference of the level energies instead.
-    # the target fractions are filled in per level by the caller, but a reader that found no
-    # photoionization data at all returns the cross-section and threshold arrays still empty, so
-    # bound the ids by what those arrays actually hold rather than indexing off the end
-    levelids_with_targets = [
-        levelid
-        for levelid, targetlist in enumerate(photoionization_targetfractions)
-        if targetlist and levelid < len(photoionization_crosssections) and levelid < len(photoionization_thresholds_ev)
-    ]
-    levelids_to_write = [
-        levelid for levelid in levelids_with_targets if not np.isnan(photoionization_thresholds_ev[levelid])
-    ]
-    skipped_no_threshold = len(levelids_with_targets) - len(levelids_to_write)
-
-    log_and_print(flog, f"Writing {len(levelids_to_write)} phixs tables to 'phixsdata2.txt'")
-    flog.write(
-        f"Downsampling cross sections assuming T={args.optimaltemperature} Kelvin, "
-        f"nphixspoints={args.nphixspoints}, phixsnuincrement={args.phixsnuincrement}\n"
-    )
-
-    if len(photoionization_crosssections) >= 1 and photoionization_crosssections[0][0] == 0.0:
-        log_and_print(flog, "ERROR: ground state has zero photoionization cross section")
-        sys.exit()
-
-    # level ids (of this ion and of the upper ion's photoionisation targets) are zero-based in
-    # memory, but the output format numbers them from one
-    for lowerlevelid in levelids_to_write:
-        targetlist = photoionization_targetfractions[lowerlevelid]
-        threshold_ev = photoionization_thresholds_ev[lowerlevelid]
-        if len(targetlist) == 1 and targetlist[0][1] > 0.99:
-            upperionlevelid = targetlist[0][0]
-
-            fphixs.write(
-                f"{atomic_number:12d}{ion_stage + 1:12d}{upperionlevelid + 1:8d}{ion_stage:12d}{lowerlevelid + 1:8d}{threshold_ev:16.6E}\n"
-            )
-        else:
-            fphixs.write(
-                f"{atomic_number:12d}{ion_stage + 1:12d}{-1:8d}{ion_stage:12d}{lowerlevelid + 1:8d}{threshold_ev:16.6E}\n"
-            )
-            fphixs.write(f"{len(targetlist):8d}\n")
-            probability_sum = 0.0
-            for upperionlevelid, targetprobability in targetlist:
-                fphixs.write(f"{upperionlevelid + 1:8d}{targetprobability:12f}\n")
-                probability_sum += targetprobability
-            if abs(probability_sum - 1.0) > 0.00001:
-                print(f"STOP! phixs fractions sum to {probability_sum:.5f} != 1.0")
-                print(targetlist)
-                print(f"level id {lowerlevelid}")
-                sys.exit()
-
-        for crosssection in photoionization_crosssections[lowerlevelid]:
-            fphixs.write(f"{crosssection:16.8E}\n")
-
-    if skipped_no_threshold > 0:
-        log_and_print(
-            flog,
-            f"Skipped {skipped_no_threshold} levels with no photoionization threshold energy",
-        )
-
-
-def write_compositionfile(
-    ion_handlers: list[tuple[int, list[int | tuple[int, str]]]], args: argparse.Namespace
-) -> None:
-    """Write compositiondata.txt, listing each element's contiguous range of ion stages."""
-    print("Writing compositiondata.txt")
-    with open(os.path.join(args.output_folder, "compositiondata.txt"), "w", encoding="utf-8") as fcomp:
-        fcomp.write(f"{len(ion_handlers):d}\n")
-        fcomp.write("0\n0\n")
-        for atomic_number, listions in ion_handlers:
-            listions_nohandlers: list[int] = drop_handlers(listions)
-            ion_stage_min: int = 0
-            ion_stage_max: int = 0
-            nions: int = 0
-            if listions_nohandlers:
-                ion_stage_min = min(listions_nohandlers)
-                ion_stage_max = max(listions_nohandlers)
-                assert all(ion_stage in listions_nohandlers for ion_stage in range(ion_stage_min, ion_stage_max + 1)), (
-                    f"Missing ion stages for Z={atomic_number} between {ion_stage_min} and {ion_stage_max}"
-                )
-                nions = ion_stage_max - ion_stage_min + 1
-
-            fcomp.write(
-                f"{atomic_number:d}  {nions:d}  {ion_stage_min:d}  {ion_stage_max:d}  "
-                f"-1 0.0 {atomic_weights[atomic_number]:.4f}\n"
-            )
-
+from artisatomic.base import atomic_weights
+from artisatomic.base import chunks
+from artisatomic.base import compression_extensions
+from artisatomic.base import elsymbols
+from artisatomic.base import empty_transitions_schema
+from artisatomic.base import find_file_check_extension
+from artisatomic.base import get_nist_ionization_energies_ev
+from artisatomic.base import h_in_ev_seconds
+from artisatomic.base import hc_in_ev_angstrom
+from artisatomic.base import hc_in_ev_cm
+from artisatomic.base import isfloat
+from artisatomic.base import leveltuples_to_pldataframe
+from artisatomic.base import log_and_print
+from artisatomic.base import parallel_map
+from artisatomic.base import path_for_log
+from artisatomic.base import PYDIR
+from artisatomic.base import roman_numerals
+from artisatomic.base import ryd_to_ev
+from artisatomic.base import split_element_ionstage_str
+from artisatomic.base import xopen_check_extension
+from artisatomic.cli import main
+from artisatomic.cli import process_files
+from artisatomic.iondata import IonData
+from artisatomic.iondata import read_ion_data
+from artisatomic.iondata import simple_handler_readers
+from artisatomic.ionhandlers import add_handler_if_not_set
+from artisatomic.ionhandlers import drop_handlers
+from artisatomic.ionhandlers import get_default_handler
+from artisatomic.ionhandlers import get_ion_handlers
+from artisatomic.ionhandlers import get_ion_stage
+from artisatomic.ionhandlers import sort_ion_handlers
+from artisatomic.levelnames import get_parity_from_config
+from artisatomic.levelnames import interpret_configuration
+from artisatomic.output import add_level_ids_forbidden
+from artisatomic.output import clear_files
+from artisatomic.output import write_adata
+from artisatomic.output import write_compositionfile
+from artisatomic.output import write_output_files
+from artisatomic.output import write_phixs_data
+from artisatomic.output import write_transition_data
+from artisatomic.phixs import match_hydrogenic_phixs
+from artisatomic.phixs import reduce_phixs_tables
+from artisatomic.phixs import reduce_phixs_tables_worker
+
+__all__ = [
+    "PYDIR",
+    "IonData",
+    "add_handler_if_not_set",
+    "add_level_ids_forbidden",
+    "atomic_weights",
+    "chunks",
+    "clear_files",
+    "compression_extensions",
+    "drop_handlers",
+    "elsymbols",
+    "empty_transitions_schema",
+    "find_file_check_extension",
+    "get_default_handler",
+    "get_ion_handlers",
+    "get_ion_stage",
+    "get_nist_ionization_energies_ev",
+    "get_parity_from_config",
+    "groundstatesonlynist",
+    "h_in_ev_seconds",
+    "hc_in_ev_angstrom",
+    "hc_in_ev_cm",
+    "interpret_configuration",
+    "isfloat",
+    "leveltuples_to_pldataframe",
+    "log_and_print",
+    "main",
+    "match_hydrogenic_phixs",
+    "parallel_map",
+    "path_for_log",
+    "process_files",
+    "read_ion_data",
+    "readboyledata",
+    "readdreamdata",
+    "readfacdata",
+    "readfloers25data",
+    "readhillierdata",
+    "readkuruczdata",
+    "readnahardata",
+    "readqubdata",
+    "readtanakajpltdata",
+    "reduce_phixs_tables",
+    "reduce_phixs_tables_worker",
+    "roman_numerals",
+    "ryd_to_ev",
+    "simple_handler_readers",
+    "sort_ion_handlers",
+    "split_element_ionstage_str",
+    "write_adata",
+    "write_compositionfile",
+    "write_output_files",
+    "write_phixs_data",
+    "write_transition_data",
+    "xopen_check_extension",
+]
 
 if __name__ == "__main__":
     main()
